feat: client feature added

+++ src/controllers/user.controller.ts
@@ -5,10 +5,18 @@ import ApiError from '../utils/ApiError';
 import { Client } from '../models/index';
 
 const createClientUser = Asyncly(async (req, res) => {
+
+	if (!req.user) {
+		throw new ApiError(httpStatus.UNAUTHORIZED, 'Authentication failed');
+	};
 	if (!req.body) {
 		throw new ApiError(httpStatus.BAD_REQUEST, 'Body cannot be empty');
 	}
-	const client = await userService.createClient(req.body);
+	const clientBody = {
+		businessOwnerId: req.user,
+		...req.body
+	}
+	const client = await userService.createClient(clientBody);
 	if (!client) {
 		throw new ApiError(httpStatus.BAD_REQUEST, 'Client not created');
 	}
@@ -20,6 +28,9 @@ const createClientUser = Asyncly(async (req, res) => {
 });
 
 const getClientUsingId = Asyncly(async (req, res) => {
+	if (!req.user) {
+		throw new ApiError(httpStatus.UNAUTHORIZED, 'Authentication failed');
+	};
 	const id = req.params?.id as string;
 	const user = await userService.getClientById(id);
 	if (!user) {
@@ -34,6 +45,9 @@ const getClientUsingId = Asyncly(async (req, res) => {
 });
 
 const getAllClient = Asyncly(async (req, res) => {
+	if (!req.user) {
+		throw new ApiError(httpStatus.UNAUTHORIZED, 'Authentication failed');
+	};
 	const client = await Client.find();
 	if (!client) {
 		throw new ApiError(httpStatus.NOT_FOUND, 'User not found');
@@ -47,6 +61,9 @@ const getAllClient = Asyncly(async (req, res) => {
 });
 
 const updateClient = Asyncly(async (req, res) => {
+	if (!req.user) {
+		throw new ApiError(httpStatus.UNAUTHORIZED, 'Authentication failed');
+	};
 	const id = req.params?.id as string;
 	const client = await userService.updateClientById(id, req.body);
 	if (!client) {
@@ -61,6 +78,9 @@ const updateClient = Asyncly(async (req, res) => {
 });
 
 const deleteClient = Asyncly(async (req, res) => {
+	if (!req.user) {
+		throw new ApiError(httpStatus.UNAUTHORIZED, 'Authentication failed');
+	};
 	const id = req.params?.id as string;
 	const client = await userService.deleteClientById(id);
 	if (!client) {
@@ -74,6 +94,9 @@ const deleteClient = Asyncly(async (req, res) => {
 });
 
 const queriedClient = Asyncly(async (req, res) => {
+	if (!req.user) {
+		throw new ApiError(httpStatus.UNAUTHORIZED, 'Authentication failed');
+	};
 	const { limit, page, where, include, exclude } = req.query;
 	const includeArray = Array.isArray(include)
 		? include.filter((item) => typeof item === 'string')
